backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
asyncpg==0.29.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
fastapi==0.104.1
httpx==0.25.2
Jinja2==3.1.2
numpy==1.26.2
pandas==2.1.4
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
psutil==5.9.6
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
scikit-learn==1.3.2
sqlalchemy==2.0.23
stripe==7.8.0
uvicorn[standard]==0.24.0
email-validator==2.1.0
firebase-admin==6.3.0
slowapi==0.1.9
